Move provider bridge into session to fix linker issues
